+++ cli/src/main.rs
@@ -54,7 +54,7 @@ fn main() -> Result<(), Box<dyn Error>> {
         ),
     };
 
-    let pretty_output = prettyplease::unparse(&syn::parse2(output).unwrap());
+    let pretty_output = prettyplease::unparse(&syn::parse_str(&output).unwrap());
 
     let output: &mut dyn Write = match &args.output_path {
         Some(path) => &mut std::fs::File::create(path).unwrap_or_else(|_| {