+++ generation/src/lir/mod.rs
@@ -1,15 +1,13 @@
 use std::ops::Range;
 
-use proc_macro2::{Ident, Literal, TokenStream};
+use crate::mir::{self, Access, BitOrder, ByteOrder, Integer};
 
-use crate::mir::{self, Access, BitOrder, ByteOrder};
-
+pub mod code_transform;
 pub mod passes;
-pub mod token_transform;
 
 pub struct Device {
-    pub internal_address_type: Ident,
-    pub register_address_type: Ident,
+    pub internal_address_type: Integer,
+    pub register_address_type: Integer,
     pub blocks: Vec<Block>,
     pub field_sets: Vec<FieldSet>,
     pub enums: Vec<Enum>,
@@ -17,19 +15,19 @@ pub struct Device {
 }
 
 pub struct Block {
-    pub cfg_attr: TokenStream,
-    pub doc_attr: TokenStream,
+    pub cfg_attr: String,
+    pub description: String,
     /// True for the root (top-level) block
     pub root: bool,
-    pub name: Ident,
+    pub name: String,
     pub methods: Vec<BlockMethod>,
 }
 
 pub struct BlockMethod {
-    pub cfg_attr: TokenStream,
-    pub doc_attr: TokenStream,
-    pub name: Ident,
-    pub address: Literal,
+    pub cfg_attr: String,
+    pub description: String,
+    pub name: String,
+    pub address: i64,
     // Only used for LIR passes, not codegen
     pub allow_address_overlap: bool,
     pub kind: BlockMethodKind,
@@ -38,35 +36,35 @@ pub struct BlockMethod {
 
 pub enum BlockMethodKind {
     Normal,
-    Repeated { count: Literal, stride: Literal },
+    Repeated { count: u64, stride: i64 },
 }
 
 pub enum BlockMethodType {
     Block {
-        name: Ident,
+        name: String,
     },
     Register {
-        field_set_name: Ident,
+        field_set_name: String,
         access: Access,
-        address_type: Ident,
-        reset_value_function: Ident,
+        address_type: Integer,
+        reset_value_function: String,
     },
     Command {
-        field_set_name_in: Option<Ident>,
-        field_set_name_out: Option<Ident>,
-        address_type: Ident,
+        field_set_name_in: Option<String>,
+        field_set_name_out: Option<String>,
+        address_type: Integer,
     },
     Buffer {
         access: Access,
-        address_type: Ident,
+        address_type: Integer,
     },
 }
 
 /// A set of fields, like a register or command in/out
 pub struct FieldSet {
-    pub cfg_attr: TokenStream,
-    pub doc_attr: TokenStream,
-    pub name: Ident,
+    pub cfg_attr: String,
+    pub description: String,
+    pub name: String,
     pub byte_order: ByteOrder,
     pub bit_order: BitOrder,
     pub size_bits: u32,
@@ -75,49 +73,65 @@ pub struct FieldSet {
     pub fields: Vec<Field>,
 }
 
+impl FieldSet {
+    pub fn size_bytes(&self) -> u32 {
+        self.size_bits.div_ceil(8)
+    }
+}
+
 pub struct Field {
-    pub cfg_attr: TokenStream,
-    pub doc_attr: TokenStream,
-    pub name: Ident,
-    pub address: Range<Literal>,
-    pub base_type: Ident,
+    pub cfg_attr: String,
+    pub description: String,
+    pub name: String,
+    pub address: Range<u32>,
+    pub base_type: String,
     pub conversion_method: FieldConversionMethod,
     pub access: mir::Access,
 }
 
 pub enum FieldConversionMethod {
     None,
-    Into(TokenStream),
-    UnsafeInto(TokenStream),
-    TryInto(TokenStream),
+    Into(String),
+    UnsafeInto(String),
+    TryInto(String),
     Bool,
 }
 
 impl FieldConversionMethod {
-    pub fn conversion_type(&self) -> Option<&TokenStream> {
+    pub fn conversion_type(&self) -> Option<&String> {
         match self {
             FieldConversionMethod::None => None,
-            FieldConversionMethod::Into(token_stream) => Some(token_stream),
-            FieldConversionMethod::UnsafeInto(token_stream) => Some(token_stream),
-            FieldConversionMethod::TryInto(token_stream) => Some(token_stream),
+            FieldConversionMethod::Into(type_path) => Some(type_path),
+            FieldConversionMethod::UnsafeInto(type_path) => Some(type_path),
+            FieldConversionMethod::TryInto(type_path) => Some(type_path),
             FieldConversionMethod::Bool => None,
         }
     }
 }
 
 pub struct Enum {
-    pub cfg_attr: TokenStream,
-    pub doc_attr: TokenStream,
-    pub name: Ident,
-    pub base_type: Ident,
+    pub cfg_attr: String,
+    pub description: String,
+    pub name: String,
+    pub base_type: String,
     pub variants: Vec<EnumVariant>,
 }
 
+impl Enum {
+    pub fn default_variant(&self) -> Option<&EnumVariant> {
+        self.variants.iter().find(|v| v.default)
+    }
+
+    pub fn catch_all_variant(&self) -> Option<&EnumVariant> {
+        self.variants.iter().find(|v| v.catch_all)
+    }
+}
+
 pub struct EnumVariant {
-    pub cfg_attr: TokenStream,
-    pub doc_attr: TokenStream,
-    pub name: Ident,
-    pub number: Literal,
+    pub cfg_attr: String,
+    pub description: String,
+    pub name: String,
+    pub number: i128,
     pub default: bool,
     pub catch_all: bool,
 }