+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - CLI: When no output file is provided, the output is printed to stdout instead
 - CLI: No longer panics when unexpected error output is processed
 - Error messages now more consistently use backticks (`) instead of various other quoting characters like (') and (")
+- Backend code generation switched from quote/syn to askama
 
 ### 1.0.4 (28-02-25)